favicons now in a separate folder
